Increased plugin compilation timeout to 2 minutes

+++ Oxide.Ext.CSharp/PluginCompiler.cs
@@ -147,12 +147,13 @@ namespace Oxide.Plugins
 
             ThreadPool.QueueUserWorkItem((_) =>
             {
-                Thread.Sleep(30000);
+                Thread.Sleep(120000);
                 if (!process.HasExited)
                 {
                     Interface.Oxide.NextTick(() =>
                     {
-                        Interface.Oxide.LogError("Timed out waiting for compiler!");
+                        if (process.HasExited) return;
+                        Interface.Oxide.LogError("Timed out waiting for compiler to compile: " + plugin.Name);
                         RemoteLogger.Error("Timed out waiting for compiler to compile: " + plugin.Name);
                         process.Kill();
                     });
